Refactor inventory data handling
